pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["tonyshh <shalin.anton8@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
gunicorn = "^21.2.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
validators = "^0.22.0"
requests = "^2.31.0"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
